pyproject.toml
[project]
name = "auto-research"
version = "0.1.0"
description = "Enterprise-grade Karpathy autoresearch for quant analysts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.50.0",
    "nbformat>=5.10.0",
    "pydantic>=2.8.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
examples = [
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "matplotlib>=3.9.0",
    "jupyter>=1.1.0",
]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages/auto_research"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["packages"]
